+++ app/v1/generic/package.scala
@@ -7,14 +7,14 @@ import v1.user.User
 import scala.concurrent.Future
 
 trait Entity {
-    val id: Long
+    val id: Int
 }
 
 package object extensions {
 
     val hashids: Hashids = Hashids.reference("p4013c7-1Ds", 7)
 
-    def decodeHash(hash: String): Option[Long] = hashids.decode(hash).headOption
+    def decodeHash(hash: String): Option[Int] = hashids.decode(hash).headOption
 
     implicit class ProjectExtensions(project: Project) {
         def hash: String = hashids.encode(project.id)
