+++ app/v1/node/NodeService.scala
@@ -6,26 +6,26 @@ import scala.concurrent.Future
 
 class NodeService @Inject()(repo: NodeRepo) {
 
-    def moveNode(nodeId: Long, move: MoveNode) = Future.successful(())
+    def moveNode(nodeId: Int, move: MoveNode) = Future.successful(())
 
-    def createNode(node: Node, parent: Option[Long]) =
+    def createNode(node: Node, parent: Option[Int]) =
         repo.createNode(node, parent)
 
-    def deleteNode(nodeId: Long) =
+    def deleteNode(nodeId: Int) =
         repo.remove(nodeId)
 
-    def renameNode(nodeId: Long, name: String) =
+    def renameNode(nodeId: Int, name: String) =
         repo.rename(nodeId, name)
 
-    def structureNodes(project: Long): Future[Seq[Node]] =
+    def structureNodes(project: Int): Future[Seq[Node]] =
         repo.findByProjectAndType(project, Structure)
 
-    def byId(nodeId: Long): Future[Option[Node]] = repo.findById(nodeId)
+    def byId(nodeId: Int): Future[Option[Node]] = repo.findById(nodeId)
 
-    def schemaNodes(project: Long): Future[Seq[Node]] =
+    def schemaNodes(project: Int): Future[Seq[Node]] =
         repo.findByProjectAndType(project, Schema)
 
-    def isChildNode(nodeId: Long, parentId: Long): Future[Boolean] = {
+    def isChildNode(nodeId: Int, parentId: Int): Future[Boolean] = {
         for {
             Some(node: Node)   <- repo.findById(nodeId)
             Some(parent: Node) <- repo.findById(parentId)
