+++ app/v1/user/User.scala
@@ -4,13 +4,13 @@ import com.mohiva.play.silhouette.api.Identity
 import v1.generic.Entity
 
 case class User(
-  id: Long,
+  id: Int,
   email: String,
   confirmed: Boolean,
   password: String,
   firstName: String,
   lastName: String,
-  lastProject: Option[Long]
+  lastProject: Option[Int]
 ) extends Identity with Entity {
     def key = email
     def fullName: String = firstName + " " + lastName
