+++ app/v1/version/VersionRepo.scala
@@ -15,7 +15,7 @@ class VersionRepo @Inject()(dcp: DatabaseConfigProvider) {
     private val dbConfig = dcp.get[PostgresProfile]
     private val db = dbConfig.db
 
-    def findById(id: Long): Future[Option[Version]] =
+    def findById(id: Int): Future[Option[Version]] =
         db.run(Versions.filter(_.id === id).result.headOption.map(_.map(toVersion)))
 
     def createVersion(version: Version, project: Project): Future[Option[Version]] =
